ParCorAna: load mask and color files as int8 and int32 respectively in case they are float. Add error checking on mask/color/finecolor, call checkParams function from defaul_config.py. Add user_params 'plot_colors' to control number of colors plotted. Add 'print_delay_curves' to get start/stop of all delay curvers on the screen
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@10344 b967ad99-d558-0410-b138-e0f6c56caec7

+++ ParCorAna/trunk/src/CommSystem.py
@@ -183,7 +183,7 @@ class MPI_Communicators:
           * workerWorldRankToCount[rank]:  number of element that worker processes
           * workerWorldRankToOffset[rank]: offset of where those elements start in 
                                            a flattened version of the ndarray
-          * maskNdarrayCoords:            the mask as a logical True/False array
+          * maskNdarrayCoords:             the mask as a logical True/False array
                                            shape has not been changed
 
         '''
@@ -1037,7 +1037,7 @@ class CommSystemFramework(object):
         mp.setLogger(verbosity)
         maskNdarrayCoords_Filename = system_params['maskNdarrayCoords']
         assert os.path.exists(maskNdarrayCoords_Filename), "mask file %s not found" % maskNdarrayCoords_Filename
-        maskNdarrayCoords = np.load(maskNdarrayCoords_Filename)
+        maskNdarrayCoords = np.load(maskNdarrayCoords_Filename).astype(np.int8)
         mp.setMask(maskNdarrayCoords)
         srcString = system_params['src']
         numEvents = system_params['numEvents']
